requirements.txt
   python-telegram-bot==13.7
   openai==0.27.0
   chromadb==0.4.18
   google-api-python-client==2.0.0
   python-dotenv==0.19.0
   numpy==1.24.3
   pandas==1.5.3
   